requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
environs==7.4.0
marshmallow==3.6.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
